Add modules and import them in index.js file
