+++ index.html
@@ -48,11 +48,10 @@
                 <li>Our addres: Streetname 22, 84503 City, Country</li>
             </ul>
         </section>
-        <footer class="footer">
-            <h3>© 2023 AwesomeBooks</h3>
-        </footer>
     </main>
-    <script src="index.js"></script>  
-    <script src="navigation.js"></script>  
+    <footer class="footer">
+        <h3>© 2023 AwesomeBooks</h3>
+    </footer>
+    <script type="module" src="./index.js"></script>  
 </body>
 </html>
